+++ supporting/completion/scoop.lua
@@ -227,7 +227,7 @@ local scoopParser = parser({
         '-q', '--quiet',
         '-h', '--help'
     ):loop(1),
-    'utils' .. parser({'auto-pr', 'checkhashes', 'checkurl', 'checkver', 'describe', 'format', 'missing-checkver'},
+    'utils' .. parser({'auto-pr', 'checkhashes', 'checkurls', 'checkver', 'describe', 'format', 'missing-checkver'},
         '--additional-options',
         '-b', '--bucketdir',
         '-h', '--help'